# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StarCash Affiliate Earning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error or invalid coupon
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /withdrawals:
    get:
      operationId: getMyWithdrawals
      tags: [withdrawals]
      summary: Get current user's withdrawal requests
      responses:
        "200":
          description: Withdrawal list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
        "401":
          description: Unauthorized
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Submit a withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: Insufficient balance or validation error
        "401":
          description: Unauthorized
...
  /admin/withdrawals:
    get:
      operationId: getAdminWithdrawals
      tags: [admin]
      summary: List all withdrawal requests
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected, completed]
          description: Filter by status
      responses:
        "200":
          description: Withdrawal list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminWithdrawal"
  /admin/withdrawals/{id}:
    patch:
      operationId: updateWithdrawal
      tags: [admin]
      summary: Approve, reject or mark a withdrawal as completed
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalStatusUpdate"
      responses:
        "200":
          description: Updated withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminWithdrawal"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, email, fullName, phone, role, status, referralCode, referralLink, walletBalance, totalEarnings, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended]
        referralCode:
          type: string
        referralLink:
          type: string
        walletBalance:
          type: number
        totalEarnings:
          type: number
        pendingEarnings:
          type: number
        withdrawableBalance:
          type: number
        totalReferrals:
          type: integer
        activeReferrals:
          type: integer
        createdAt:
          type: string
          format: date-time
    RegisterInput:
      type: object
      required: [fullName, username, email, phone, password, confirmPassword, couponCode]
      properties:
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
          minLength: 6
        confirmPassword:
          type: string
        referralCode:
          type: ["string", "null"]
        couponCode:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
...
    Withdrawal:
      type: object
      required: [id, userId, bankName, accountName, accountNumber, amount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        bankName:
          type: string
        accountName:
          type: string
        accountNumber:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected, completed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WithdrawalInput:
      type: object
      required: [bankName, accountName, accountNumber, amount]
      properties:
        bankName:
          type: string
        accountName:
          type: string
        accountNumber:
          type: string
        amount:
          type: number
          minimum: 6
...
    WithdrawalStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected, completed]
    AdminWithdrawal:
      type: object
      required: [id, userId, username, fullName, bankName, accountName, accountNumber, amount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        fullName:
          type: string
        bankName:
          type: string
        accountName:
          type: string
        accountNumber:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected, completed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...