# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * StarCash Affiliate Earning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  username: string;
  email: string;
  fullName: string;
  phone: string;
  role: UserRole;
  status: UserStatus;
  referralCode: string;
  referralLink: string;
  walletBalance: number;
  totalEarnings: number;
  pendingEarnings?: number;
  withdrawableBalance?: number;
  totalReferrals?: number;
  activeReferrals?: number;
  createdAt: Date;
}